feat: MySQL 고정 및 Validation 오류 제거

+++ src/main/java/com/team/saver/common/util/InitData.java
@@ -270,6 +270,11 @@ public class InitData implements CommandLineRunner {
     @Override
     @Transactional
     public void run(String... args) throws Exception {
+        if(accountRepository.findAll().size() != 0) {
+
+            return;
+        }
+
         deleteAllObjectsInBucket();
         marketDocumentRepository.deleteAll();
